+++ .vscode/settings.json
@@ -9,5 +9,8 @@
   },
   "tslint.enable": true,
   "tslint.autoFixOnSave": true,
-  "typescript.tsdk": "node_modules/typescript/lib"
+  "typescript.tsdk": "node_modules/typescript/lib",
+  "editor.codeActionsOnSave": {
+    "source.organizeImports": "explicit"
+  },
 }
