+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "serverless-simple-middleware",
   "description": "Simple middleware to translate the interface of lambda's handler to request => response",
-  "version": "0.0.63",
+  "version": "0.0.64",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "author": "VoyagerX",
@@ -22,12 +22,12 @@
   "homepage": "https://github.com/v6x/serverless-simple-middleware#readme",
   "scripts": {
     "lint": "tslint --project tsconfig.json",
+    "clean": "rm -rf dist && rm -rf node_modules && yarn install",
     "build": "tsc --project tsconfig.json",
     "build:watch": "tsc --project tsconfig.json --watch",
     "test": "jest --config jest.config.js",
     "test:watch": "jest --config jest.config.js --watch",
-    "deploy": "yarn build && yarn publish",
-    "clean": "rm -rf dist && rm -rf node_modules && yarn install"
+    "deploy": "yarn build && yarn publish"
   },
   "dependencies": {
     "@aws-sdk/client-dynamodb": "^3.828.0",
@@ -40,7 +40,9 @@
     "@types/aws-lambda": "8",
     "@types/mysql": "^2.15.5",
     "cross-fetch": "^2.2.2",
+    "kysely": "^0.28.2",
     "mysql": "^2.16.0",
+    "mysql2": "^3.14.1",
     "nanoid": "4.0.2",
     "p-limit": "^2.0.0",
     "simple-staging": "^0.0.12",
