+++ src/aws/index.ts
@@ -1,3 +1,3 @@
-export * from './define';
 export * from './config';
+export * from './define';
 export * from './simple';
