+++ src/middleware/database/index.ts
@@ -1,0 +1,10 @@
+export { ConnectionProxy } from './connectionProxy';
+export {
+  sql,
+  SQLClient,
+  type DeleteQueryBuilder,
+  type ExpressionBuilder,
+  type InsertQueryBuilder,
+  type SelectQueryBuilder,
+  type UpdateQueryBuilder,
+} from './sqlClient';
