+++ tsconfig.json
@@ -1,8 +1,8 @@
 {
   "compilerOptions": {
-    "target": "es5",
+    "target": "ES2022",
     "module": "commonjs",
-    "lib": ["es2017", "es7", "es6", "esnext", "dom"],
+    "lib": ["ES2022", "es7", "es6", "esnext", "dom"],
     "rootDir": "src",
     "outDir": "dist",
     "declaration": true,
